requirements.txt
python-telegram-bot==20.7
anthropic==0.23.1
PyPDF2==3.0.1
python-docx==1.1.0
